PkgDependency.jl

Quick Start

julia> import PkgDependency
julia> PkgDependency.tree("Tables")
 Tables v1.10.0
━━━━━━━━━━━━━━━━
       │
       ├── DataAPI v1.13.0
       ├── OrderedCollections v1.4.1
       ├── IteratorInterfaceExtensions v1.0.0
       ├── DataValueInterfaces v1.0.0
       └── TableTraits v1.0.1
           └── IteratorInterfaceExtensions v1.0.0 (*)

PkgDependency.tree() can be used to print dependency tree of current workspace.

Unless otherwise specified, all methods of tree function support following kwargs:

kwarg	default	description

reverse	false	get a reverse dependency tree

compat	false	show compat info in tree

show_link	false	show packages' repo link in tree

dedup	true	hide duplicate dependencies in tree

stdlib	false	show packages from Standard Library

API

PkgDependency.tree — Function

tree(...; reverse=false, compat=false, show_link=false, dedup=true, stdlib=false)

PkgDependency.tree — Method

tree(name::AbstractString; kwargs...)

Print dependency tree of a package identified by name.

PkgDependency.tree — Method

tree(uuid::UUID; kwargs...)

Print dependency tree of a package identified by UUID

PkgDependency.tree — Method

tree(; kwargs...)

Print dependency tree of current project. reverse kwarg is not supported in this method.